backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
pydantic
geopy
python-multipart
python-dotenv
